Cleaned up game.py and ui.py, creating a UI class for better defined objects and easier function calling.
